+++ packages/@ourworldindata/components/src/index.ts
@@ -27,3 +27,19 @@ export {
 export { ExpandableToggle } from "./ExpandableToggle/ExpandableToggle.js"
 
 export { IndicatorBrief } from "./IndicatorBrief/IndicatorBrief.js"
+
+export {
+    IndicatorSources,
+    type OriginSubset,
+} from "./IndicatorSources/IndicatorSources.js"
+
+export {
+    CodeSnippet,
+    hydrateCodeSnippets,
+    renderCodeSnippets,
+} from "./CodeSnippet/CodeSnippet.js"
+
+export {
+    DATAPAGE_SOURCES_AND_PROCESSING_SECTION_ID,
+    REUSE_THIS_WORK_SECTION_ID,
+} from "./SharedDataPageConstants.js"
