+++ site/owid.scss
@@ -70,7 +70,7 @@
 @import "./blocks/CookiePreferences.scss";
 @import "./blocks/Grid.scss";
 @import "./blocks/Card.scss";
-@import "./blocks/code-snippet.scss";
+@import "../packages/@ourworldindata/components/src/CodeSnippet/code-snippet.scss";
 @import "./blocks/BiographyCard.scss";
 @import "./blocks/KeyInsights.scss";
 @import "./blocks/TechnicalText.scss";
